+++ examples/model_generator.py
@@ -1,3 +1,8 @@
+"""
+An example script for cobea using randomly generated Response objects.
+
+Bernard Riemann (bernard.riemann@tu-dortmund.de
+"""
 from numpy import abs
 from numpy.random import randint, randn, uniform
 from cobea.model import BEModel, Topology, Response
@@ -43,7 +48,7 @@ def random_model(K, J, M, include_dispersion, max_180deg=True):
 
 
 def random_response_drift(K, J, M=1, include_dispersion=True, relative_noise=0.02,
-                          make_drift=True, hidden_file='hidden_model.pickle'):
+                          make_drift=True, hidden_filename=None):
     """generate a random, valid Response object.
 
     Parameters
@@ -90,34 +95,43 @@ def random_response_drift(K, J, M=1, include_dispersion=True, relative_noise=0.0
     # now just make some noise...
     measured_matrix += relative_noise * randn(K, J, M)
 
-    with open(hidden_file, 'wb') as f:
-        dump(hidden_model, f, protocol=2)
-        print('hidden model saved to ' + hidden_file)
+    if hidden_filename is not None:
+        with open(hidden_filename, 'wb') as f:
+            dump(hidden_model, f, protocol=2)
+            print('hidden model saved to ' + hidden_filename)
 
     return Response(measured_matrix, topology.corr_names, topology.mon_names,
                     topology.line, include_dispersion, unit='a.u.'), drift
 
 
 if __name__ == '__main__':
-    from cobea import cobea, normalize_using_drift
+    from cobea import cobea
+    from cobea.pproc import normalize_using_drift
     import cobea.plotting as plt
 
-    response, drift_space = random_response_drift(30, 32)
+    hidden_filename = 'hidden_model.pickle'
+    result_filename = None # 'cobea_result.pickle'
+
+    response, drift_space = random_response_drift(30, 32, hidden_filename=hidden_filename)
 
     result = cobea(response, drift_space)
 
+    if result_filename is not None:
+        result.save(result_filename)
+
+
     # plot results in comparison with hidden model
-    with open('hidden_model.pickle', 'rb') as f:
+    with open(hidden_filename, 'rb') as f:
         hidden_model = load(f)
     hidden_data = {'name': 'hidden', 'beta_jmw': hidden_model.cbeta_jmw,
                    'delphi_jmw': hidden_model.delphi_jmw, 'd_jw': hidden_model.d_jw,
                    'b_k': hidden_model.b_k, 'beta_km': hidden_model.cbeta_km,
                    'delphi_km': hidden_model.delphi_km}
 
-    save_path = 'model_generator_output/'
+    pdf_path = 'model_generator_output/'
     try:
-        makedirs(save_path)
+        makedirs(pdf_path)
     except OSError:
         pass  # directory already exists
 
-    plt.plot_result(result, prefix=save_path, comparison_data=hidden_data)
+    plt.plot_result(result, prefix=pdf_path, comparison_data=hidden_data)